refac: rename package for member service
